+++ src/mixins/Allocations.lua
@@ -2,16 +2,39 @@
 local Addon = select(2, ...)
 local Util = Addon.Util
 
----@type ProfessionTransationAllocations?
-local AllocationsMixin
+---@class Allocations.Static
+---@field Mixin ProfessionTransationAllocations
+local Self = Addon.Allocations
 
 ---@return ProfessionTransationAllocations
-function Addon:CreateAllocations()
-    if not AllocationsMixin then
+function Self:Create()
+    if not Self.Mixin then
         local recipe = C_TradeSkillUI.GetRecipeSchematic(898, false)
         local transaction = CreateProfessionsRecipeTransaction(recipe)
-        AllocationsMixin = Util:TblCreateMixin(transaction:GetAllocations(1))
+        Self.Mixin = Util:TblCreateMixin(transaction:GetAllocations(1))
     end
 
-    return CreateAndInitFromMixin(AllocationsMixin)
+    local allocs = Mixin({}, Self.Mixin)
+    self:WithoutOnChanged(allocs, "Init")
+
+    return allocs
+end
+
+---@param allocs ProfessionTransationAllocations
+---@param fn function | string
+---@param ... any[]
+function Self:WithoutOnChanged(allocs, fn, ...)
+    local origOnChanged = allocs.OnChanged
+    allocs.OnChanged = Util.FnNoop
+
+    local res
+    if type(fn) == "function" then
+        res = fn(...)
+    else
+        res = allocs[fn](allocs, ...)
+    end
+
+    allocs.OnChanged = origOnChanged
+
+    return res
 end